src: ban nondeterministic dispatch, retire the last format-string appender
Ban 22: virtual functions, class hierarchies, RTTI and std::function. A virtual
call reads its target out of the object, so the worst-case path is not in the
image - the same nondeterminism the fixed pools and compile-time sizing exist to
remove. src/ was already clean of all four; the gate keeps it that way, proven by
injecting each form and watching it report at the exact line.

Retire fmtbuf.h. Its three JSON emitters (dashboard, gpio_map, partition_monitor)
now declare pc_field specs and call the frame engine, so nothing re-parses a
format string at runtime to rediscover what the spec already states. Labels and
keys go through PC_FK_JSON, which escapes them - the old %s did not, so a widget
label carrying a quote produced invalid JSON.

sftp longname drops its snprintf and its staging buffer: pc_sb_u64_clip takes the
column width directly, which is what the ls -l alignment actually is.

Strict typing in hex.h, http_date.h and the three serializers - size_t hides a
ceiling the config already fixed and is a different width on host and target.
pc_hex_u32 counts its digits first and writes each to its final index, so it needs
no reversal buffer.

+++ src/services/storage/partition_monitor/partition_monitor.h
@@ -49,7 +49,7 @@ const char *pc_partition_kind(uint8_t type, uint8_t subtype);
  * @brief Serialize a partition array as JSON `{"partitions":[...]}` into @p out.
  * @return characters written, or 0 if @p cap is too small.
  */
-int pc_partition_json(const pc_partition_info *parts, uint8_t count, char *out, size_t cap);
+int32_t pc_partition_json(const pc_partition_info *parts, uint8_t count, char *out, uint32_t cap);
 
 /**
  * @brief Walk the flash partition table into @p out (ESP32; 0 on host builds).
